fix dubbo-samples-generic
