+++ dubbo-samples-generic/dubbo-samples-generic-call/src/main/java/org/apache/dubbo/samples/generic/call/GenericCallConsumer.java
