+++ dubbo-samples-generic/dubbo-samples-generic-call/src/main/resources/log4j.properties
