+++ dubbo-samples-generic/dubbo-samples-generic-call/src/main/resources/spring/generic-provider.xml
