+++ dubbo-samples-generic/dubbo-samples-generic-impl/src/main/java/org/apache/dubbo/samples/generic/impl/GenericImplProvider.java
@@ -1,0 +1,54 @@
+/*
+ *
+ *   Licensed to the Apache Software Foundation (ASF) under one or more
+ *   contributor license agreements.  See the NOTICE file distributed with
+ *   this work for additional information regarding copyright ownership.
+ *   The ASF licenses this file to You under the Apache License, Version 2.0
+ *   (the "License"); you may not use this file except in compliance with
+ *   the License.  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package org.apache.dubbo.samples.generic.impl;
+
+import org.apache.dubbo.config.ApplicationConfig;
+import org.apache.dubbo.config.RegistryConfig;
+import org.apache.dubbo.config.ServiceConfig;
+import org.apache.dubbo.rpc.service.GenericService;
+import org.apache.dubbo.samples.generic.impl.impl.GenericImplOfHelloService;
+
+import java.util.concurrent.CountDownLatch;
+
+
+public class GenericImplProvider {
+    private static String zookeeperAddress = "zookeeper://" + System.getProperty("zookeeper.address", "127.0.0.1") + ":2181";
+
+    public static void main(String[] args) throws Exception {
+        new EmbeddedZooKeeper(2181, false).start();
+
+        ApplicationConfig applicationConfig = new ApplicationConfig();
+        applicationConfig.setName("generic-impl-provider");
+        RegistryConfig registryConfig = new RegistryConfig();
+        registryConfig.setAddress(zookeeperAddress);
+
+        GenericService helloService = new GenericImplOfHelloService();
+        ServiceConfig<GenericService> service = new ServiceConfig<>();
+        service.setApplication(applicationConfig);
+        service.setRegistry(registryConfig);
+        service.setInterface("org.apache.dubbo.samples.generic.impl.api.HelloService");
+        service.setRef(helloService);
+        service.setGeneric("true");
+        service.export();
+
+        System.out.println("dubbo service started");
+        new CountDownLatch(1).await();
+    }
+}
