+++ dubbo-samples-generic/dubbo-samples-generic-impl/src/main/resources/log4j.properties
