+++ dubbo-samples-generic/dubbo-samples-generic-type/src/main/resources/spring/generic-type-consumer.xml
@@ -27,7 +27,7 @@
 
     <dubbo:application name="generic-type-consumer"/>
 
-    <dubbo:registry address="zookeeper://${zookeeper.address:127.0.0.1}:2181"/>
+    <dubbo:registry address="zookeeper://${zookeeper.address:127.0.0.1}:${zookeeper.port:2181}"/>
 
     <dubbo:reference id="userService" interface="org.apache.dubbo.samples.generic.api.UserService"/>
 </beans>
